Revert "trying to fix auto scroll on iOS safari"
This reverts commit 95b3d26daf6adc4fd149a4fc217878e6df98e679.

+++ src/components/tracker/TrackerItem.jsx
@@ -52,6 +52,7 @@ function TrackerItem({children, item, data, itemIndex}) {
 
   const [touchEnded, setTouchEnded] = useState(false)
   const [deleting, setDeleting] = useState(false)
+  const [pinning, setPinning] = useState(false)
   const [assessmentOptions, setAssessmentOptions] = useState(false)
   const [selectedDays, setSelectedDays] = useState(item.reminderDays)
 
@@ -78,7 +79,7 @@ function TrackerItem({children, item, data, itemIndex}) {
   function onScroll(e) {
     itemExtended.current = e.target.scrollLeft === e.target.children[0].clientWidth || e.target.clientWidth + e.target.children[0].clientWidth*2
 
-    if (e.target.scrollLeft > e.target.scrollWidth - e.target.clientWidth*1.5 && touchEnded && !deleting) {
+    if (e.target.scrollLeft > e.target.scrollWidth - e.target.clientWidth*1.5 && touchEnded && !deleting && !pinning) {
       deleteItem()
     }
   }
@@ -147,10 +148,17 @@ function TrackerItem({children, item, data, itemIndex}) {
 
 
   function pinItem() {
+    setPinning(true)
     const updatedItems = getSortedItems(items.map(i => i.id === item.id ? {...i, pinned: !i.pinned} : {...i}),assessments)
 
     setItems(updatedItems)
     saveItems(updatedItems)
+
+    scrollEl.current.scrollTo({left: scrollEl.current.children[0].clientWidth, behavior: 'instant'})
+
+    setTimeout(() => {
+      setPinning(false)
+    }, 500)
   }
 
   function onSetPinned() {
@@ -240,7 +248,6 @@ function TrackerItem({children, item, data, itemIndex}) {
       }
     }
     itemContainer.current.classList.remove('hiding-animation')
-    scrollEl.current.scrollTo({left: scrollEl.current.children[0].clientWidth, behavior: 'instant'})
   }, [assessments, item, item.reminderDays, deleting])
 
   const assessmentProps = {
